specify arch specific nanorc

+++ modules/040-nano/install.sh
@@ -3,9 +3,11 @@ source $DOT_ROOT/constants.sh
 DOT_MODULE="nano"
 
 log info applying nanorc
-log silly backing up nanorc
-[ -e "$HOME/.nanorc" ] && [ ! -e "$HOME/.nanorc.bak" ] && cp "$HOME/.nanorc" "$HOME/.nanorc.bak"
 
-log silly linking nanorc
-rm "$HOME/.nanorc"
-ln -s "$DOT_ROOT/modules/040-nano/.nanorc" "$HOME/.nanorc"
+if [ "$DOT_OS" = "linux_arch" ]; then
+    log silly backing up nanorc
+    [ -e "$HOME/.nanorc" ] && [ ! -e "$HOME/.nanorc.bak" ] && cp "$HOME/.nanorc" "$HOME/.nanorc.bak"
+    log silly linking nanorc
+    rm "$HOME/.nanorc"
+    ln -s "$DOT_ROOT/modules/040-nano/.nanorc" "$HOME/.nanorc"
+fi
